Refactor wrapper/container
