+++ packages/venia-concept/src/components/Checkout/index.js
@@ -1,1 +1,1 @@
-export { default } from './wrapper';
+export { default } from './container';
